Collect Default Model Drivers

Collect Default Model Drivers

calcDefaultDrivers(drivers = c("Population", "GDP", "UrbanPop"))

Arguments

drivers

Vector of strings.

Value

A magpie object with sets "iso3c", "year" and "variable".

Return supplementary information

Set the supplementary argument of madrat::calcOutput() to TRUE to return a list with the scenarios and additional information on the unit, the data sources and the harmonization function.

Vectorization of arguments

Vectors are accepted for -Calib, -Past and -Future arguments. If given a vector, different combinations are created and returned all at once. If more than one argument is vectorised, the arguments have to have the same length. Which time series are created can be illustrated with the following example. Let's say the harmonization function and past data source are vectors of length 3. Then there will be in total 3 time series that are produced: the first time series is the result of combining the first harmonization function with the first past data source, the second time series the result of combining the second harmonization function with the second past data source, and the third time series the result of using the respective third entry. The future data source used in each case is the same, since in this example only one future data source is provided.

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("DefaultDrivers")
}