Get GDP scenarios and building blocks

Get complete GDP scenarios with calcGDP, or the past/future scenario building blocks with calcGDPPast and calcGDPFuture.

Complete scenarios are created by harmonizing future projections (returned by calcGDPFuture) onto historical data (returned by calcGDPPast) and cover the years between 1960 and 2100.

If GDP data for a scenario is required, even if just for a single year, always use calcGDP, as what is returned by calcGDPPast or calcGDPFuture may not end up as is in the scenario, depending on the harmonization function used (see the GDPcalib argument for more information). Use calcGDPPast and calcGDPFuture only when trying to access specific GDP data, or when constructing new complete scenarios.

By default, calcGDP returns the following scenarios:

the SSPs, i.e. SSP1-5 and SSP2EU

the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC

Usage

calcGDP(
  GDPCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  GDPPast = c("WDI-MI", "WDI-MI", "Eurostat-WDI-MI"),
  GDPFuture = c("SSPs-MI", "SDPs-MI", "SSP2EU-MI"),
  unit = "constant 2005 Int$PPP",
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  average2020 = TRUE,
  naming = "indicator_scenario"
)

calcGDPFuture(
  GDPFuture = "SSPs-MI",
  unit = "constant 2005 Int$PPP",
  extension2150 = "none"
)

calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2005 Int$PPP")

Arguments

GDPCalib

A string designating the harmonization function. Available harmonization functions are:

"calibSSPs": use past data, short term growth rates from IMF and afterwards transition between GDPPast and GDPFuture with a transition period until 2100

"calibSSP2EU": use past data, short term growth rates from IMF and afterwards transition between GDPPast and GDPFuture with a transition period until 2100. For European countries, just glue past with future and after 2070 converge to 2150 SSP2 values.

"calibSDPs": same as calibSSPs

"past": deprecated

"future": deprecated

"transition": deprecated

"past_transition": use past data and afterwards transition between GDPPast and GDPFuture with a transition period until 2050

GDPPast

A string designating the source for the historical GDP data. Available sources are:

"WDI": World development indicators from the World Bank

"MI": Missing island dataset

"Eurostat": Eurostat

"PWT": Penn World Tables

The IHME/James data set. This data set is tertiary source that uses available secondary sources (e.g. WDI and PWT) to create completed GDP per capita time series (See paper). To access, pass one of the subtypes "IHME_USD05_PPP_pc", "IHME_USD05_MER_pc", "IMF_USD05_PPP_pc", "PENN_USD05_PPP_pc", "WB_USD05_PPP_pc", "MADDISON_USD05_PPP_pc", "WB_USD05_MER_pc", "IMF_USD05_MER_pc", "UN_USD05_MER_pc". In all cases, the per capita GDP will be multiplied by WDI population data to get absoulte GDP data.

See the "Combining data sources with '-'" section below for how to combine data sources.

GDPFuture

A string designating the source for the future GDP data. Available sources are:

"SSPs": IIASA SSP database

"SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source

"SDPs":

"MI": Missing island dataset

"OECD": OECD

"SRES": deprecated

See the "Combining data sources with '-'" section below for how to combine data sources.

unit

A string specifying the unit of GDP. Can be either:

"constant 2005 Int$PPP" (default): Default unit in which the scenarios are constructed.

"constant 2005 US$MER": Scenarios are constructed in constant 2005 Int$PPP and converted before being returned with GDPuc::convertGDP.

"constant 2017 Int$PPP": Scenarios are constructed in constant 2017 Int$PPP (See paper).

"constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and converted before being returned with GDPuc::convertGDP.

In all cases, GDP is returned in millions.

extension2150

A string specifying if/how the scenarios should be extended until 2150. Can be either:

"bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as their last year.

"constant": The last value of the scenarios is taken as constant until 2150.

"none": No extension.

FiveYearSteps

FiveYearSteps = TRUE will be deprecated soon. Until then, please set FiveYearSteps = FALSE and use the calcOutput argument years instead, to retrieve specific years.

average2020

TRUE (default) or FALSE. If TRUE, then the 2020 value is replaced by the 2018-2022 average.

naming

A string giving the naming scheme of the data dimension. Can be either:

"indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".

"indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".

"scenario": Returns names of the type "SSP2".

Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to worry about the conflicting names.

Value

A magpie object with sets "iso3c", "year" and "variable".

Functions

calcGDPFuture: Get future GDP projections

calcGDPPast: Get historic GDP data

Combining data sources with "-"

Past and Future data sources can be combined with "-" and given as arguments to the function, i.e. "WDI-MI". This signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.

Vectorization of arguments

Vectors are accepted for -Calib, -Past and -Future arguments. If given a vector, different combinations are created and returned all at once. If more than one argument is vectorised, the arguments have to have the same length. Which time series are created can be illustrated with the following example. Let's say the harmonization function and past data source are vectors of length 3. Then there will be in total 3 time series that are produced: the first time series is the result of combining the first harmonization function with the first past data source, the second time series the result of combining the second harmonization function with the second past data source, and the third time series the result of using the respective third entry. The future data source used in each case is the same, since in this example only one future data source is provided.

Return supplementary information

Set the supplementary argument of madrat::calcOutput() to TRUE to return a list with the scenarios and additional information on the unit, the data sources and the harmonization function.

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("GDP")
# or, for the now-outdated GDP scenarios used before summer 2021,
calcOutput("GDP",
           GDPCalib = "past_transition",
           GDPPast = "IHME_USD05_PPP_pc-MI",
           GDPFuture = "SSPs-MI",
           extension2150 = "none",
           average2020 = FALSE,
           aggregate = FALSE,
           FiveYearSteps = FALSE)
}

if (FALSE) {
library(mrdrivers)
calcOutput("GDPFuture")
}

if (FALSE) {
library(mrdrivers)
calcOutput("GDPPast")
}