Get GDP and GDPpc scenario building blocks

Get the past and future GDP scenario building blocks with calcGDPPast and calcGDPFuture, respectively. If GDP data for a scenario is required, even if just for a single year, always use calcGDP(), as what is returned by calcGDPPast or calcGDPFuture may not end up as is in the scenario, depending on the harmonization function. Use calcGDPPast and calcGDPFuture only when trying to access specific GDP data.

See the "Combining data sources with '-'" section below for how to combine data sources.

Usage

calcGDPFuture(GDPFuture, unit)

calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2017 Int$PPP")

calcGDPpcFuture(GDPpcFuture, unit)

calcGDPpcPast(GDPpcPast = "WDI-MI", unit = "constant 2017 Int$PPP")

Arguments

GDPFuture

A string designating the source for the future GDP data. Available sources are:

"SSPs": IIASA SSP database

"SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source

"SDPs":

"MI": Missing island dataset

See the "Combining data sources with '-'" section below for how to combine data sources.

unit

A string specifying the unit of GDP. Can be either:

"constant 2017 Int$PPP" (default): Scenarios are constructed in constant 2017 Int$PPP.

"constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and then converted with GDPuc::toolConvertGDP().

In all cases, GDP is returned in millions.

GDPPast

A string designating the source for the historical GDP data. Available sources are:

"WDI": World development indicators from the World Bank

"MI": Missing island dataset

"Eurostat": Eurostat

GDPpcFuture

A string designating the source for the future GDP data. Available sources are:

"SSPs":

"SDPs": All SDP future projections are set equal to those of SSP1.

"MI": Missing island dataset

See the "Combining data sources with '-'" section below for how to combine data sources.

GDPpcPast

A string designating the source for the historical GDPpc data. Available sources are:

"WDI": World development indicators from the World Bank

"MI": Missing island dataset

See the "Combining data sources with '-'" section below for how to combine data sources.

Functions

calcGDPFuture(): Get future GDP projections

calcGDPpcFuture(): Get future GDPpc projections

calcGDPpcPast(): Get historic GDPpc data

Combining data sources with "-"

Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.