calcPopulation, calcPopulationPast, calcPopulationFuture

Get complete population scenarios with calcPopulation, or the past/future scenario building blocks with calcPopulationPast and calcPopulationFuture.

Complete scenarios are created by harmonizing future projections (returned by calcPopulationFuture) onto historical data (returned by calcPopulationPast) and cover the years between 1960 and 2100.

If population data for a scenario is required, even if just for a single year, always use calcPopulation, as what is returned by calcPopulationPast or calcPopulationFuture may not end up as is in the scenario, depending on the harmonization function used (see the Populationcalib argument for more information). Use calcPopulationPast and calcPopulationFuture only when trying to access specific population data, or when constructing new complete scenarios.

By default, calcPopulation returns the following scenarios:

the SSPs, i.e. SSP1-5 and SSP2EU

the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC

calcPopulation(
  PopulationCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  PopulationPast = c("WDI-UN_PopDiv-MI", "WDI-UN_PopDiv-MI",
    "Eurostat-WDI-UN_PopDiv-MI"),
  PopulationFuture = c("SSPs-UN_PopDiv-MI", "SDPs-UN_PopDiv-MI", "SSP2EU-UN_PopDiv-MI"),
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)

calcPopulationFuture(
  PopulationFuture = "SSPs-UN_PopDiv-MI",
  extension2150 = "none"
)

calcPopulationPast(PopulationPast = "WDI-UN_PopDiv-MI")

Arguments

PopulationCalib

A string designating the harmonization function. Available harmonization functions are:

"calibSSPs": use past data from PopulationPast, then the growth rates from the Wolrld Bank's PEAP until the final year of the IMF WEO data, and then the growth rates from PopulationFuture.

"calibSSP2EU":

"calibSDPs": same as calibSSPs

"past": deprecated

"future": deprecated

"transition": deprecated

"past_transition":

"past_grFuture":

PopulationPast

A string designating the source for the historical population data. Available sources are:

"WDI": World development indicators from the World Bank

"UN_PopDiv": United Nations

"MI": Missing island dataset

"Eurostat": Eurostat

See the "Combining data sources with '-'" section below for how to combine data sources.

PopulationFuture

A string designating the source for the future population data. Available sources are:

"SSPs": From the Wittgenstein Center here and here

"SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source

"SDPs":

"UN_PopDiv": United Nations

"MI": Missing island dataset

"SSPs_old": Old SSPs from the IIASA database

"SRES": deprecated

"IIASApop": deprecated

See the "Combining data sources with '-'" section below for how to combine data sources.

extension2150

A string specifying if/how the scenarios should be extended until 2150. Can be either:

"bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as their last year.

"constant": The last value of the scenarios is taken as constant until 2150.

"none": No extension.

FiveYearSteps

FiveYearSteps = TRUE is no longer supported; use the calcOutput argument years instead, to retrieve specific years.

naming

A string giving the naming scheme of the data dimension. Can be either:

"indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".

"indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".

"scenario": Returns names of the type "SSP2".

Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to worry about the conflicting names.

Value

A magpie object with sets "iso3c", "year" and "variable".

Functions

calcPopulationFuture: Get future population projections

calcPopulationPast: Get historic population data

Combining data sources with "-"

Past and Future data sources can be combined with "-" and given as arguments to the function, i.e. "WDI-MI". This signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.

Vectorization of arguments

Vectors are accepted for -Calib, -Past and -Future arguments. If given a vector, different combinations are created and returned all at once. If more than one argument is vectorised, the arguments have to have the same length. Which time series are created can be illustrated with the following example. Let's say the harmonization function and past data source are vectors of length 3. Then there will be in total 3 time series that are produced: the first time series is the result of combining the first harmonization function with the first past data source, the second time series the result of combining the second harmonization function with the second past data source, and the third time series the result of using the respective third entry. The future data source used in each case is the same, since in this example only one future data source is provided.

Return supplementary information

Set the supplementary argument of madrat::calcOutput() to TRUE to return a list with the scenarios and additional information on the unit, the data sources and the harmonization function.

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("GDPpc")
}

if (FALSE) {
library(mrdrivers)
calcOutput("PopulationFuture")
}

if (FALSE) {
library(mrdrivers)
calcOutput("PopulationPast")
}