Get Population and Labour scenario building blocks

See the "Combining data sources with '-'" section below for how to combine data sources.

Usage

calcLabourFuture(LabourFuture = "SSPs")

calcLabourPast(LabourPast = "WDI")

calcPopulationFuture(PopulationFuture = "SSPs-UN_PopDiv-MI")

calcPopulationPast(PopulationPast = "WDI-UN_PopDiv-MI")

calcUrbanFuture(UrbanFuture = "SSPs")

calcUrbanPast(UrbanPast = "WDI")

Arguments

LabourFuture

A string designating the source for the future working-age population data. Available sources are:

"SSPs":

LabourPast

A string designating the source for the historical working-age population data. Available sources are:

"WDI": World development indicators from the World Bank

PopulationFuture

A string designating the source for the future population data. Available sources are:

"SSPs": From the Wittgenstein Center here and here

"SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source

"SDPs":

"UN_PopDiv": United Nations

"MI": Missing island dataset

PopulationPast

A string designating the source for the historical population data. Available sources are:

"WDI": World development indicators from the World Bank

"UN_PopDiv": United Nations

"MI": Missing island dataset

"Eurostat": Eurostat

UrbanFuture

A string designating the source for the future urban population-share data. Available sources are:

"SSPs":

"SDPs":

"SSP2EU":

UrbanPast

A string designating the source for the historical urban population-share data. Available sources are:

"WDI": World development indicators from the World Bank

Combining data sources with "-"

Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.