calcUrban

Get urbanization scenarios. By default the following scenarios are returned:

the SSPs, i.e. SSP1-5 and SSP2EU

the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC

The scenarios are created by harmonizing future projections onto historical data.

calcUrban(
  UrbanCalib = "past",
  UrbanPast = "WDI",
  UrbanFuture = c("SSPs", "SDPs", "SSP2EU"),
  extension2150 = "constant",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)

Arguments

UrbanCalib

To what should be calibrated? past or future?

UrbanPast

Urban past data source

UrbanFuture

Urban future data source

extension2150

A string specifying if/how the scenarios should be extended until 2150. Can be either:

"bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as their last year.

"constant": The last value of the scenarios is taken as constant until 2150.

"none": No extension.

FiveYearSteps

FiveYearSteps = TRUE will be deprecated soon. Until then, please set FiveYearSteps = FALSE and use the calcOutput argument years instead, to retrieve specific years.

naming

A string giving the naming scheme of the data dimension. Can be either:

"indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".

"indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".

"scenario": Returns names of the type "SSP2".

Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to worry about the conflicting names.

Value

A magpie object with sets "iso3c", "year" and "variable".

Return supplementary information

Set the supplementary argument of madrat::calcOutput() to TRUE to return a list with the scenarios and additional information on the unit, the data sources and the harmonization function.

Vectorization of arguments

Vectors are accepted for -Calib, -Past and -Future arguments. If given a vector, different combinations are created and returned all at once. If more than one argument is vectorised, the arguments have to have the same length. Which time series are created can be illustrated with the following example. Let's say the harmonization function and past data source are vectors of length 3. Then there will be in total 3 time series that are produced: the first time series is the result of combining the first harmonization function with the first past data source, the second time series the result of combining the second harmonization function with the second past data source, and the third time series the result of using the respective third entry. The future data source used in each case is the same, since in this example only one future data source is provided.

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("Urban")
}