Extend until 2150 in 5 year time steps

toolExtend2150 extends a magpit object until 2150 in 5 year time steps. Either with bezierExtension or constant.

Usage

toolExtend2150(data, extension2150)

Arguments

data

A list with "x", a magpie object, and "description" elements

extension2150

A string specifying if/how the scenarios should be extended until 2150. Can be either:

"bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as their last year.

"constant": The last value of the scenarios is taken as constant until 2150.

"none": No extension.

Value

The modified data list.