Get information on available scenarios

toolGetScenarioDefinition can be used to figure out which scenarios are made available by mrdrivers, and how they are constructed, i.e. what xPast, xFuture and xCalib strings are used.

Usage

toolGetScenarioDefinition(driver = NULL, scen = NULL, unlist = FALSE)

Arguments

driver

NULL or a character vector designating the driver for which information is to be returned. If NULL, information for all drivers is returned. Available drivers are "GDP", "Population" and "GDPpc".

scen

NULL or a character vector designating the scenario for which information is to be returned. If NULL, information for all scenarios is returned. Available scenarios are:

GDP-scenarios: "SSPs", "SDPs", and "SSP2EU"

Populations-scenarios: "SSPs", "SDPs", "SSP2EU" and "ISIMIP"

GDPpc-scenarios: "SSPs", "SDPs", and "SSP2EU"

unlist

TRUE or FALSE (default). If TRUE the xCalib, xPast and xFuture entries are returned directly.

Value

A list with the scenarios and their xPast, xFuture and xCalib strings.

Examples

toolGetScenarioDefinition()
#> $GDP
#> # A tibble: 7 × 4
#>   scenario   GDPPast              GDPFuture GDPCalib       
#>   <chr>      <chr>                <chr>     <chr>          
#> 1 SSPsOld    IHME_USD05_PPP_pc-MI SSPs-MI   past_transition
#> 2 SSPs       WDI-MI               SSPs-MI   calibSSPs      
#> 3 SDPs       WDI-MI               SDPs-MI   calibSDPs      
#> 4 SSP2EU     Eurostat-WDI-MI      SSP2EU-MI calibSSP2EU    
#> 5 noCovid    WDI-MI               SSPs-MI   calibNoCovid   
#> 6 longCovid  WDI-MI               SSPs-MI   calibLongCovid 
#> 7 shortCovid WDI-MI               SSPs-MI   calibShortCovid
#> 
#> $Population
#> # A tibble: 5 × 4
#>   scenario PopulationPast            PopulationFuture    PopulationCalib
#>   <chr>    <chr>                     <chr>               <chr>          
#> 1 SSPs     WDI-UN_PopDiv-MI          SSPs-UN_PopDiv-MI   calibSSPs      
#> 2 SDPs     WDI-UN_PopDiv-MI          SDPs-UN_PopDiv-MI   calibSDPs      
#> 3 SSP2EU   Eurostat-WDI-UN_PopDiv-MI SSP2EU-UN_PopDiv-MI calibSSP2EU    
#> 4 ISIMIP   UN_PopDiv-MI              SSPs-UN_PopDiv-MI   calibISIMIP    
#> 5 SSPsOld  WDI-MI                    SSPs_old-MI         past_transition
#> 
#> $GDPpc
#> # A tibble: 6 × 4
#>   scenario   GDPpcPast       GDPpcFuture GDPpcCalib     
#>   <chr>      <chr>           <chr>       <chr>          
#> 1 SSPs       WDI-MI          SSPs-MI     calibSSPs      
#> 2 SDPs       WDI-MI          SDPs-MI     calibSDPs      
#> 3 SSP2EU     Eurostat-WDI-MI SSP2EU-MI   calibSSP2EU    
#> 4 noCovid    WDI-MI          SSPs-MI     calibNoCovid   
#> 5 longCovid  WDI-MI          SSPs-MI     calibLongCovid 
#> 6 shortCovid WDI-MI          SSPs-MI     calibShortCovid
#> 
toolGetScenarioDefinition(driver = "GDP")
#> $GDP
#> # A tibble: 7 × 4
#>   scenario   GDPPast              GDPFuture GDPCalib       
#>   <chr>      <chr>                <chr>     <chr>          
#> 1 SSPsOld    IHME_USD05_PPP_pc-MI SSPs-MI   past_transition
#> 2 SSPs       WDI-MI               SSPs-MI   calibSSPs      
#> 3 SDPs       WDI-MI               SDPs-MI   calibSDPs      
#> 4 SSP2EU     Eurostat-WDI-MI      SSP2EU-MI calibSSP2EU    
#> 5 noCovid    WDI-MI               SSPs-MI   calibNoCovid   
#> 6 longCovid  WDI-MI               SSPs-MI   calibLongCovid 
#> 7 shortCovid WDI-MI               SSPs-MI   calibShortCovid
#> 
toolGetScenarioDefinition(scen = "SSP2EU")
#> $GDP
#> # A tibble: 1 × 4
#>   scenario GDPPast         GDPFuture GDPCalib   
#>   <chr>    <chr>           <chr>     <chr>      
#> 1 SSP2EU   Eurostat-WDI-MI SSP2EU-MI calibSSP2EU
#> 
#> $Population
#> # A tibble: 1 × 4
#>   scenario PopulationPast            PopulationFuture    PopulationCalib
#>   <chr>    <chr>                     <chr>               <chr>          
#> 1 SSP2EU   Eurostat-WDI-UN_PopDiv-MI SSP2EU-UN_PopDiv-MI calibSSP2EU    
#> 
#> $GDPpc
#> # A tibble: 1 × 4
#>   scenario GDPpcPast       GDPpcFuture GDPpcCalib 
#>   <chr>    <chr>           <chr>       <chr>      
#> 1 SSP2EU   Eurostat-WDI-MI SSP2EU-MI   calibSSP2EU
#> 
toolGetScenarioDefinition(driver = "Population", scen = "SSPs", unlist = TRUE)
#> $PopulationPast
#> [1] "WDI-UN_PopDiv-MI"
#> 
#> $PopulationFuture
#> [1] "SSPs-UN_PopDiv-MI"
#> 
#> $PopulationCalib
#> [1] "calibSSPs"
#>