Get GDP and GDP per capita scenarios

Like all scenarios in mrdrivers, the GDP and GDP per capita scenarios are the result of a harmonization exercise between past data and future projections. Together with the corresponding population scenarios (see calcPopulation()) they comprise a consistent set of scenarios.

By default the following scenarios are returned:

the SSPs, i.e. SSP1-5

the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC

See the vignette: vignette("scenarios") for scenario options, definitions and references.

Usage

calcGDP(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)

calcGDPpc(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)

Arguments

scenario

A string (or vector of strings) designating the scenario(s) to be returned. Use toolGetScenarioDefinition() to learn what scenarios are available.

unit

A string specifying the unit of GDP. Can be either:

"constant 2017 Int$PPP" (default): Scenarios are constructed in constant 2017 Int$PPP.

"constant 2017 US$MER": Scenarios are constructed in constant 2017 US$MER and then converted with GDPuc::toolConvertGDP().

In all cases, GDP is returned in millions.

average2020

If TRUE (default), then the 2020 value is replaced by the 2018-2022 average. To be consistent, the yearly resolution is decreased to 5 year intervals.

...

Arguments passed on to calcDriver(), of which "extension2150" and "naming" are most often of interest.

Value

magpie object with the requested output data either on country or on regional level depending on the choice of argument "aggregate" or a list of information if supplementary is set to TRUE.

See also

Examples

if (FALSE) { # \dontrun{
# Return default scenarios
calcOutput("GDP")
calcOutput("GDPpc")

# Return only the SSP2 GDP scenario
calcOutput("GDP", scenario = "SSP2")
} # }

if (FALSE) { # \dontrun{
calcOutput("GDPpc")
} # }