calcGDPPast

Calculates a time series of GDP in Purchase Power Parity (PPP) of million International Dollars of the year 2005. Different sources are available:

WDI: The PPP estimate from the World Development Indicators (WDI) are supplemented by values for Argentina, Syria and Somalia which are missing in the database. The values were taken from World Bank. 2014. Purchasing Power Parities and the Real Size of World Economies: A Comprehensive Report of the 2011 International Comparison Program. The World Bank. http://elibrary.worldbank.org/doi/book/10.1596/978-1-4648-0329-1. table 2.13 Then, the 2011 estimate is extrapolated with the GDP projection in local currency units (LCU), as these growth rates should be approximately what the growth models predict. The price index from 2011 was transformed into 2005 equivalents using the inflation rate of the United States (US), as the PPPs are in USDollar equvialents.

PWT: Penn World Tables

IHME_USD05_PPP_pc: Publication: James, Spencer L., Paul Gubbins, Christopher JL Murray, and Emmanuela Gakidou. 2012. "Developing a Comprehensive Time Series of GDP per Capita for 210 Countries from 1950 to 2015."" Population Health Metrics 10 (1): 12. doi:10.1186/1478-7954-10-12. This publication also contains further interpolated indicators that can be used.

calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2005 Int$PPP")

Arguments

GDPPast
GDP past data source

unit
A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER', 'constant 2017 Int$PPP' or 'constant 2017 US$MER'. In either case, the scenario construction is done in Int$PPP, with the conversion to US$MER coming afterward.

Value

A magpie object with sets "iso3c", "year" and "variable".

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("GDPPast")
}