Get scenario building blocks

Get the past and future scenario building blocks. If scenario data is required, even if just for a single year, always use the calc call to the final scenario, e.g. calcGDP() or calcPopulation(), as what is returned by the calc...Past and calc...Future functions may not end up as is in the scenario, depending on the harmonization function. These functions are only still exported for compatibility with existing code in the PIAM input data pipeline.

See the vignette: vignette("scenarios") for references for the different data sources.

See the "Combining data sources with '-'" section below for how to combine data sources.

Usage

calcGDPPast(
  pastData = toolGetScenarioDefinition("GDPpc", "SSPs")$pastData,
  extension1960 = "MI-James"
)

calcGDPpcPast(scenario = "SSPs")

calcGDPFuture(futureData)

calcGDPpcFuture(scenario)

calcPopulationFuture(futureData)

calcLabourFuture(futureData)

calcUrbanFuture(futureData)

calcPopulationPast(
  pastData = toolGetScenarioDefinition("Population", "SSPs")$pastData
)

calcLabourPast(pastData = toolGetScenarioDefinition("Labour", "SSPs")$pastData)

calcUrbanPast(pastData = toolGetScenarioDefinition("Urban", "SSPs")$pastData)

Arguments

pastData

A string specifying the sources for historic data.

extension1960

A string specifying how the data should be extended backwards to 1960.

scenario

A string specifying the scenario, from which the historic GDP and population data sources are taken.

futureData

A string specifying the sources for future projections.

Combining data sources with "-"

Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.