calcPopulation

Create population time series by harmonizing future projections onto historical data.

calcPopulation(
  PopulationCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  PopulationPast = c("WDI-UN_PopDiv-MI", "WDI-UN_PopDiv-MI",
    "Eurostat-WDI-UN_PopDiv-MI"),
  PopulationFuture = c("SSPs-UN_PopDiv-MI", "SDPs-UN_PopDiv-MI", "SSP2EU-UN_PopDiv-MI"),
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)

Arguments

PopulationCalib
A string designating the harmonization function. Available harmonization functions are:

"calibSSPs":

"calibSSP2EU":

"calibSDPs":

"calibUN_PopDiv":

"past": deprecated

"future": deprecated

"transition": deprecated

"past_transition":

"past_grFuture":

PopulationPast
A string designating the source for the historical population data. Available sources are:

"WDI": World development indicators from the World Bank

"UN_PopDiv": United Nations

"MI": Missing island dataset

"Eurostat": Eurostat

PopulationFuture
A string designating the source for the future population data. Available sources are:

"SSPs":

"SSP2EU":

"SDPs":

"UN_PopDiv":

"MI":

"SSPs_old":

"SRES": deprecated

"IIASApop": deprecated

extension2150
string, either "bezier", "constant" or "none"

FiveYearSteps
FiveYearSteps = TRUE is no longer supported; use the calcOutput argument years instead, to retrieve specific years.

naming
naming scheme

Value

A magpie object with sets "iso3c", "year" and "variable".

Details

The function accepts vectors for the harmonization function and past and future data sources. If given a vector, different combinations are created and returned all at once. If more than one argument is vectorised, the arguments have to have the same length. Which time series are created can be illustrated with the following example. Let's sy the harmonization function and past data source are vetors of length 3. Then there will be in total 3 time series that are produced: the first time series is the result of combining the first hamonization function with the first past data source, the second time series the result of combining the second hamonization function with the second past data source, and the third time series the result of using the repective third entry. The future data source used in each case is the same, since in this example only one future data source is provided.

See also

Examples

if (FALSE) {
library(mrdrivers)
calcOutput("GDPpc")
}