Get population and labour scenarios

Like all scenarios in mrdrivers, the Population, Labour and Urban population share scenarios are the result of a harmonization exercise between past data and future projections.

By default the following scenarios are returned:

the SSPs, i.e. SSP1-5

the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC

See the vignette: vignette("scenarios") for scenario options, definitions and references.

Usage

calcPopulation(scenario = c("SSPs", "SDPs", "SSP2EU"), ...)

calcLabour(scenario = c("SSPs", "SDPs", "SSP2EU"), ...)

calcUrban(scenario = c("SSPs", "SDPs", "SSP2EU"), asShare = TRUE, ...)

Arguments

scenario

A string (or vector of strings) designating the scenario(s) to be returned. Use toolGetScenarioDefinition() to learn what scenarios are available.

...

Arguments passed on to calcDriver(), of which "extension2150" and "naming" are most often of interest. Other calcDriver() arguments are used for scenario fine-tuning and by package developers.

asShare

If TRUE (default) urban population shares are returned. If FALSE, then urban population in millions is returned.

Value

magpie object with the requested output data either on country or on regional level depending on the choice of argument "aggregate" or a list of information if supplementary is set to TRUE.

See also

Examples

if (FALSE) { # \dontrun{
# Return the default scenarios
calcOutput("Population")

# Return only the SSP2 scenario
calcOutput("Population", scenario = "SSP2")

# Return the ISIMIP SSP scenarios
calcOutput("Population", scenario = "ISIMIP", extension2150 = "none", aggregate = FALSE)
} # }
if (FALSE) { # \dontrun{
calcOutput("Labour")
} # }
if (FALSE) { # \dontrun{
calcOutput("Urban")
} # }