Get information on available scenarios

toolGetScenarioDefinition can be used to figure out which scenarios are made available by mrdrivers, and how they are constructed, i.e. what past data, future data and harmonization methods are used.

Usage

toolGetScenarioDefinition(driver = NULL, scen = NULL, aslist = FALSE)

Arguments

driver

NULL or a character vector designating the driver for which information is to be returned. If NULL, information for all drivers is returned. Available drivers are:

GDP

Population

GDPpc

Labour

Urban

scen

NULL or a character vector designating the scenario for which information is to be returned. If NULL, information for all scenarios is returned.

aslist

TRUE or FALSE (default). If TRUE then the pastData, futureData and harmonization strings are returned as a list.

Value

A tibble with the driver and scenario information.

Examples

toolGetScenarioDefinition()
#> # A tibble: 31 × 5
#>    driver scenario      pastData     futureData       harmonization     
#>    <chr>  <chr>         <chr>        <chr>            <chr>             
#>  1 GDPpc  SSPs          WDI-MI-James SSPs             GDPpcSSPs         
#>  2 GDPpc  SSP2          WDI-MI-James SSP2             GDPpcSSPs         
#>  3 GDPpc  SSP2EU        WDI-MI-James SSP2EU           GDPpcSSPs         
#>  4 GDPpc  SDPs          -            -                GDPpcSDPs         
#>  5 GDPpc  ISIMIP        WDI-MI-James SSPs             GDPpcSSPs         
#>  6 GDPpc  SSP2IndiaDEAs WDI-MI-James IndiaDEAs-SSP2   GDPpcSSP2IndiaDEAs
#>  7 GDPpc  SSP2IndiaHigh WDI-MI-James IndiaDEAopt-SSP2 GDPpcSSP2IndiaDEAs
#>  8 GDP    SSPs          -            -                GDPpcWithPop      
#>  9 GDP    SSP2          -            -                GDPpcWithPop      
#> 10 GDP    SSP2EU        -            -                GDPpcWithPop      
#> # ℹ 21 more rows
toolGetScenarioDefinition(driver = "GDP")
#> # A tibble: 7 × 5
#>   driver scenario      pastData futureData harmonization
#>   <chr>  <chr>         <chr>    <chr>      <chr>        
#> 1 GDP    SSPs          -        -          GDPpcWithPop 
#> 2 GDP    SSP2          -        -          GDPpcWithPop 
#> 3 GDP    SSP2EU        -        -          GDPpcWithPop 
#> 4 GDP    SDPs          -        -          GDPpcWithPop 
#> 5 GDP    ISIMIP        -        -          GDPpcWithPop 
#> 6 GDP    SSP2IndiaDEAs -        -          GDPpcWithPop 
#> 7 GDP    SSP2IndiaHigh -        -          GDPpcWithPop 
toolGetScenarioDefinition(scen = "SSP2")
#> # A tibble: 5 × 5
#>   driver     scenario pastData         futureData     harmonization
#>   <chr>      <chr>    <chr>            <chr>          <chr>        
#> 1 GDPpc      SSP2     WDI-MI-James     SSP2           GDPpcSSPs    
#> 2 GDP        SSP2     -                -              GDPpcWithPop 
#> 3 Population SSP2     WDI-UN_PopDiv-MI SSP2-UN_PopDiv PopSSPs      
#> 4 Labour     SSP2     WDI-UN_PopDiv    SSP2-UN_PopDiv pastAndLevel 
#> 5 Urban      SSP2     WDI              SSP2           pastAndGrowth
toolGetScenarioDefinition(driver = "Population", scen = "SSPs", aslist = TRUE)
#> $driver
#> [1] "Population"
#> 
#> $scenario
#> [1] "SSPs"
#> 
#> $pastData
#> [1] "WDI-UN_PopDiv-MI"
#> 
#> $futureData
#> [1] "SSPs-UN_PopDiv"
#> 
#> $harmonization
#> [1] "PopSSPs"
#>